move logic to cleanup.go
